requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
click-plugins==1.1.1
cligj==0.7.2
cloud-init==21.2
cloudpickle==1.6.0
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextvars==2.4
cryptography==2.1.4
Cython==0.29.24
dask==2021.3.0
distributed==2021.3.0
distro-info===0.18ubuntu0.18.04.1
et-xmlfile==1.1.0
Fiona==1.8.20
fsspec==2021.7.0
GeoAlchemy2==0.9.2
geopandas==0.9.0
greenlet==1.1.0
HeapDict==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
immutables==0.15
importlib-metadata==4.6.1
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
locket==0.2.1
MarkupSafe==1.0
msgpack==1.0.2
munch==2.5.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==2.0.6
openpyxl==3.0.7
PAM==0.4.2
pandas==1.1.5
partd==1.2.0
pexpect==4.2.1
psutil==5.8.0
psycopg2==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyproj==3.0.1
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
Rtree==0.9.7
SecretStorage==2.3.1
service-identity==16.0.0
Shapely==1.7.1
six==1.16.0
sortedcontainers==2.4.0
sos==4.1
soupsieve==2.2.1
SQLAlchemy==1.4.22
ssh-import-id==5.7
systemd-python==234
tblib==1.7.0
toolz==0.11.1
tornado==6.1
Twisted==17.9.0
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
WALinuxAgent==2.2.45
xlrd==2.0.1
zict==2.0.0
zipp==3.5.0
zope.interface==4.3.2
